requirements.txt
matplotlib
numpy
tensorflow-gpu
